// frontend/fingertips-frontend/generated-sources/ft-api-client/models/Frequency.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Fingertips API
 * An API to query public health indicator data.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: ProfileFeedback@dhsc.gov.uk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * a type of frequency eg Annual, monthly.
 * @export
 */
export const Frequency = {
    Annual: 'Annual',
    Monthly: 'Monthly',
    Quarterly: 'Quarterly',
    CumulativeQuarters: 'CumulativeQuarters'
} as const;
export type Frequency = typeof Frequency[keyof typeof Frequency];


export function instanceOfFrequency(value: any): boolean {
    for (const key in Frequency) {
        if (Object.prototype.hasOwnProperty.call(Frequency, key)) {
            if (Frequency[key as keyof typeof Frequency] === value) {
                return true;
            }
        }
    }
    return false;
}

export function FrequencyFromJSON(json: any): Frequency {
    return FrequencyFromJSONTyped(json, false);
}

export function FrequencyFromJSONTyped(json: any, ignoreDiscriminator: boolean): Frequency {
    return json as Frequency;
}

export function FrequencyToJSON(value?: Frequency | null): any {
    return value as any;
}

export function FrequencyToJSONTyped(value: any, ignoreDiscriminator: boolean): Frequency {
    return value as Frequency;
}